Docker: update target database to latest PostGIS version
This ensures that source and target database PostGIS are the same
version. Otherwise it could happen that the Docker image version is
pinned by the base image to a version lower than the latest while the
source database is update to the latest PostGIS version (according to
apt). This can break the PostGIS update if the updated source version
introduces functionality that isn't available in the (not upgraded)
target database. This is what happened with the 13/3.2 Docker image,
when PostGIS 3.3 was already released. The newer version introduced a
new function (pgis_geometry_union_parallel_combinefn), which broke the
pg_upgrade, because the new function wasn't available in the target
database (running PostGIS 3.2). With this change, the target database is
also updated to the most recent PostGIS version.